step-by-step/step2/pyproject.toml
[project]
name = "simple-agent-step2"
version = "0.1.0"
description = "Simple Python agent using Microsoft 365 Agent SDK - Step 2 (Azure Bot Service)"
requires-python = ">=3.10"
dependencies = [
    "microsoft-agents-hosting-aiohttp>=0.6.1",
    "microsoft-agents-authentication-msal>=0.6.1",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0"
]

[tool.uv]
dev-dependencies = []
